Rename GDPR component, Privacy Dialog and decouple the actual card from privacy Dialog.
