+++ src/components/gdpr/GdprSettingsCard.vue
